# apps/backend/src/app.ts
import express, { type Application } from 'express';
import { createServer, type Server } from 'http';
import { Server as SocketServer } from 'socket.io';
import cors from 'cors';
import helmet from 'helmet';
import compression from 'compression';
import { config } from './config/index.js';
import { authRouter } from './routes/auth.routes.js';
import { websiteRouter } from './routes/website.routes.js';
import { analyzerRouter } from './routes/analyzer.routes.js';
import { openapiRouter } from './routes/openapi.routes.js';
import { historyRouter } from './routes/history.routes.js';
import { compareHistoryRouter } from './routes/compareHistory.routes.js';
import { adviceRouter }         from './routes/advice.routes.js';
import { authAuditRouter } from './routes/authAudit.routes.js';
import { competitorSessionRouter } from './routes/competitorSession.routes.js';
import { onboardingRouter } from './routes/onboarding.routes.js';
import { overviewRouter }   from './routes/overview.routes.js';
import { notificationsRouter } from './routes/notifications.routes.js';
import { cliAuthRouter }           from './routes/cliAuth.routes.js';
import { rumRouter } from './routes/rum.routes.js';
import { deployRouter } from './routes/deploy.routes.js';
import { flowRouter } from './routes/flow.routes.js';
import { teamRouter } from './routes/team.routes.js';
import { cruxRouter }              from './routes/crux.routes.js';
import { registerAnalysisSocket } from './socket/analysis.handler.js';
import { registerFlowSocket } from './socket/flow.handler.js';
import { markStorageState, STORAGE_HEADER } from './middleware/storage.middleware.js';
import { requestLog } from './middleware/requestLog.middleware.js';
import { attachTeamScope, TEAM_HEADER } from './middleware/teamScope.js';
import { isDbReady } from './config/database.js';
import { errorMiddleware } from './lib/errors.js';
import { renderMetrics } from './lib/metrics.js';
import { registerRuntimeCollectors } from './lib/runtimeMetrics.js';
import type { ServerToClientEvents, ClientToServerEvents } from '@perfscope/shared';
import type { InterServerEvents, SocketData } from './types/socket.js';

/**
 * How long a request may block before the server hangs up.
 *
 * Deliberately short, and deliberately *not* raised to match `RUN_TIMEOUT_MS` (4 min) in
 * lighthouse.service.ts: this is server-wide, so a value big enough for an audit would let
 * every wedged request hold a socket for four minutes. The one route that legitimately runs
 * longer raises the limit for its own connection — see analyzer.routes.ts.
 */
const HTTP_TIMEOUT_MS = 70_000;

export function createApp(): { app: Application; httpServer: Server } {
  registerRuntimeCollectors();

  const app = express();
  const httpServer = createServer(app);

  httpServer.setTimeout(HTTP_TIMEOUT_MS);

  // Whose address `req.ip` is. Behind the deployment's nginx the real client is the last
  // entry of X-Forwarded-For, and Express will not read it without being told how many
  // hops to trust — see config.trustProxy for why it is a count and not `true`.
  app.set('trust proxy', config.trustProxy);

  // ── Socket.io ────────────────────────────────────────────────────────────
  const io = new SocketServer<
    ClientToServerEvents,
    ServerToClientEvents,
    InterServerEvents,
    SocketData
  >(httpServer, {
    cors: { origin: config.clientUrl, methods: ['GET', 'POST'] },
  });

  registerAnalysisSocket(io);
  registerFlowSocket(io);

  // ── Middleware ───────────────────────────────────────────────────────────
  /**
   * The request log, and the id every line written under it carries.
   *
   * First in the chain on purpose: a request rejected by CORS or answered by a 404 is
   * still a request somebody made, and those are exactly the ones a report is about.
   */
  app.use(requestLog);

  /**
   * Security headers.
   *
   * `crossOriginResourcePolicy` is explicitly cross-origin, and that is not a formality:
   * helmet's default is `same-origin`, and this server serves `/rum.js` to be loaded as a
   * script *by other people's sites*. The default would have turned the RUM snippet into a
   * blocked request on every site that installed it — and the failure would have shown up
   * as "field data shows nothing", the same symptom as an unset key.
   */
  app.use(helmet({ crossOriginResourcePolicy: { policy: 'cross-origin' } }));

  /**
   * Compression, mounted before the routes so every JSON response goes through it. An
   * audit result is hundreds of kilobytes of audits, waterfall rows and treemap nodes, and
   * it is fetched again every time somebody reopens a stored run.
   */
  app.use(compression());

  // Two headers that only work if they are named. `exposedHeaders`: a browser cannot read
  // a custom *response* header across origins unless the server exposes it. `allowedHeaders`:
  // naming any request header at all replaces cors' reflect-whatever-was-asked default, so
  // the three the client already sent have to be listed again beside the team one.
  app.use(cors({
    origin: config.clientUrl,
    exposedHeaders: [STORAGE_HEADER],
    allowedHeaders: ['Content-Type', 'Authorization', 'X-Requested-With', ...(config.teamsEnabled ? [TEAM_HEADER] : [])],
  }));
  app.use(express.json());
  app.use(markStorageState);

  /**
   * Liveness and readiness in one, above the routers so nothing can gate it.
   *
   * There was no health route at all, which meant every check of "is the backend up" was
   * somebody loading a page — and a container platform has no page to load. `status` is
   * about the process; `database` is reported separately and deliberately does NOT make
   * this 503: the app is designed to serve empty shapes with no Mongo (see
   * storage.middleware), so a database outage is a degradation to report, not a reason to
   * have the orchestrator restart a server that is working as designed.
   */
  app.get('/health', (_req, res) => {
    res.json({
      status:    'ok',
      uptime:    Math.round(process.uptime()),
      database:  isDbReady() ? 'up' : 'down',
      // Which build is answering. `APP_VERSION` is baked into the image at release time
      // (a tag, or `main-<sha>`), and is the only way to tell from outside whether a
      // deploy actually replaced the running container.
      version:   config.appVersion,
    });
  });

  /**
   * Metrics, for whatever scrapes them.
   *
   * **Not proxied by the dashboard's nginx** — it forwards `/api`, `/socket.io`, `/rum.js`
   * and `/health`, and nothing else — so on the standard deployment this is reachable only
   * from inside the compose network. `METRICS_TOKEN` is there for an install that does
   * expose it; unset, the endpoint is open, which is correct exactly as long as the port
   * is not.
   */
  app.get('/metrics', (req, res) => {
    if (config.metricsToken && req.headers.authorization !== `Bearer ${config.metricsToken}`) {
      res.status(401).type('text/plain').send('unauthorized\n');
      return;
    }
    res.type('text/plain; version=0.0.4').send(renderMetrics());
  });

  // ── Routes ───────────────────────────────────────────────────────────────
  // Mounted at the root: it serves /rum.js as well as /api/rum, and carries its own
  // permissive CORS because both are called from other origins.
  app.use(rumRouter);

  /**
   * Which account a request acts on, resolved once for every router below.
   *
   * Above the routers rather than inside one: they all share the bare `/api` mount, so a
   * `router.use()` here would leak into every router registered after it — the trap that
   * once broke CLI login. It runs before `requireAuth` and hands it `scopeUserId`.
   *
   * Both the middleware and the router exist only with `TEAMS_ENABLED=true`: with the
   * feature off, every request is the person's own and `/api/teams` is a 404.
   */
  if (config.teamsEnabled) app.use('/api', attachTeamScope);

  app.use('/api', authRouter);
  if (config.teamsEnabled) app.use('/api', teamRouter);
  app.use('/api', websiteRouter);
  app.use('/api', deployRouter);
  app.use('/api', flowRouter);
  app.use('/api', analyzerRouter);
  app.use('/api', openapiRouter);
  app.use('/api', historyRouter);
  app.use('/api', compareHistoryRouter);
  app.use('/api', adviceRouter);
  app.use('/api', authAuditRouter);
  app.use('/api', competitorSessionRouter);
  app.use('/api', cruxRouter);
  app.use('/api', onboardingRouter);
  app.use('/api', overviewRouter);
  app.use('/api', notificationsRouter);
  app.use('/api/auth', cliAuthRouter);

  app.get('/', (_req, res) => {
    res.json({ name: 'PerfScope API', version: '1.0.0', status: 'running' });
  });

  app.use((_req, res) => {
    res.status(404).json({ success: false, error: 'Route not found' });
  });

  // Last in the chain, after the routes: anything a handler throws lands here and
  // becomes a response. Without it an uncaught throw hung until the 70s timeout.
  app.use(errorMiddleware);

  return { app, httpServer };
}