# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PerfScope API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0",
    "description": "Every route the backend serves, what it is for, and whether it needs a token. Kept honest by\n`apps/backend/src/routes/routes.contract.test.ts`, which fails when a route is added, moved\nor removed without this file changing.\n\n**Machine-readable:** [`openapi.json`](openapi.json) (OpenAPI 3.1) is generated from these\ntables and the `@perfscope/shared` types — run `pnpm openapi` in `apps/backend` after changing\neither, and `apps/backend/openapi/operations.ts` says which type each route answers with. CI\nfails on a stale spec, and `e2e/openapi-conformance.probe.mjs` validates real responses\nagainst it.\n\nGenerated from `docs/api/README.md` and the `@perfscope/shared` types by `apps/backend/openapi/build.mts` — edit those, never this file."
  },
  "servers": [
    {
      "url": "/",
      "description": "The origin serving this PerfScope install"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/rum.js": {
      "get": {
        "tags": [
          "Real-user monitoring"
        ],
        "summary": "The collector script, served at the root and loaded from other people's sites",
        "description": "The collector script, served at the **root** and loaded from other people's sites. Cacheable — it is the same file for every site.",
        "operationId": "getRumJs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/javascript": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": []
      }
    },
...
    "/api/teams": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Teams this person owns or belongs to",
        "description": "Teams this person owns or belongs to.",
        "operationId": "getTeams",
        "parameters": [
          {
            "$ref": "#/components/parameters/TeamId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TeamSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Create one",
        "description": "Create one.",
        "operationId": "postTeams",
        "parameters": [
          {
            "$ref": "#/components/parameters/TeamId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/TeamDetail"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from /api/auth/login, /register or /refresh. Lasts 30 minutes."
      }
    },
    "parameters": {
      "TeamId": {
        "name": "X-Team-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Act inside a team: the request is resolved to the team owner's account."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid or expired access token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error": {
        "description": "Any failure: `{ success: false, error }`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "TeamDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TeamRole",
            "description": "The caller's own role, not the team's. A list is read to decide where to work."
          },
          "members": {
            "type": "number"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "memberList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamMemberInfo"
            }
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "id",
          "memberList",
          "members",
          "name",
          "ownerId",
          "ownerName",
          "role"
        ],
        "description": "One team with its people."
      },
...
      "TeamMemberInfo": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TeamRole"
          },
          "joinedAt": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name",
          "email",
          "role",
          "joinedAt"
        ],
        "description": "A person on a team, as the members list shows them."
      },
      "TeamRole": {
        "type": "string",
        "enum": [
          "viewer",
          "member",
          "owner"
        ]
      },
      "TeamSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TeamRole",
            "description": "The caller's own role, not the team's. A list is read to decide where to work."
          },
          "members": {
            "type": "number"
          },
          "ownerName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "members",
          "ownerName"
        ],
        "description": "A team in a list — enough to switch to it and to say what you can do there."
      },
...
    }
  }
}